Add files via upload
This is a final coursework project on Database design in my first year of the data science program at the Southern Alberta Institute of Technology. In the data landscape, one of the most sought-after technical skills is the ability to efficiently develop a scalable database design for your clients and customers while being a consultant. I delved into the best practices for a database design model to build an efficient database for "Crocs" (referred to as my client). 

An outline of how a database can be efficiently designed from the ground up to satisfy changing business requirements will be given in this article. My name is Kehinde, and I am a technical writer/Data and data science enthusiast. 

Let's connect more: [LinkedIn](https://www.linkedin.com/in/adeniran-olanrewaju/)

Please click [here]() to read the entire article. Thank you.